refactor(tests): share OTLP log parse-back helper across telemetry tests
Addresses a review note on PR #980: index.test.ts's inline attrsOf only read stringValue while handlers/workflows.test.ts had its own attrText handling intValue/doubleValue — two copies of the same parse-back logic. Both files now use logEvents() from otlp-test-utils.ts.
